// frontend/src/api/store/store.ts
export type getInventoryResponse200 = {
  data: GetInventory200
  status: 200
}

export type getInventoryResponseDefault = {
  data: void
  status: Exclude<HTTPStatusCodes, 200>
}

export type getInventoryResponseSuccess = (getInventoryResponse200) & {
  headers: Headers;
};
export type getInventoryResponseError = (getInventoryResponseDefault) & {
  headers: Headers;
};

export type getInventoryResponse = (getInventoryResponseSuccess | getInventoryResponseError)