// frontend/src/api/pet/pet.ts
export type findPetsByStatusResponse200ApplicationJson = {
  data: Pet[]
  status: 200
}

export type findPetsByStatusResponse200ApplicationXml = {
  data: Pet[]
  status: 200
}

export type findPetsByStatusResponse400 = {
  data: void
  status: 400
}

export type findPetsByStatusResponseDefault = {
  data: void
  status: Exclude<HTTPStatusCodes, 200 | 400>
}

export type findPetsByStatusResponseSuccess = (findPetsByStatusResponse200ApplicationJson | findPetsByStatusResponse200ApplicationXml) & {
  headers: Headers;
};
export type findPetsByStatusResponseError = (findPetsByStatusResponse400 | findPetsByStatusResponseDefault) & {
  headers: Headers;
};

export type findPetsByStatusResponse = (findPetsByStatusResponseSuccess | findPetsByStatusResponseError)